added maven support created pom.xml, added maven version placeholders to plugin.yml, changed api-version to 1.13, set jdk to 1.8
