--- src/ru/alskar/signmap/PersistentUUID.java
@@ -1,40 +1,0 @@
-package ru.alskar.signmap;
-
-import org.bukkit.persistence.PersistentDataAdapterContext;
-import org.bukkit.persistence.PersistentDataType;
-
-import java.nio.ByteBuffer;
-import java.util.UUID;
-
-// Thanks for the class implementation and the great guide to:
-// https://www.spigotmc.org/threads/a-guide-to-1-14-persistentdataholder-api.371200/
-// It was of a great help!
-
-public class PersistentUUID implements PersistentDataType<byte[], UUID> {
-
-    @Override
-    public Class<byte[]> getPrimitiveType() {
-        return byte[].class;
-    }
-
-    @Override
-    public Class<UUID> getComplexType() {
-        return UUID.class;
-    }
-
-    @Override
-    public byte[] toPrimitive(UUID complex, PersistentDataAdapterContext context) {
-        ByteBuffer bb = ByteBuffer.wrap(new byte[16]);
-        bb.putLong(complex.getMostSignificantBits());
-        bb.putLong(complex.getLeastSignificantBits());
-        return bb.array();
-    }
-
-    @Override
-    public UUID fromPrimitive(byte[] primitive, PersistentDataAdapterContext context) {
-        ByteBuffer bb = ByteBuffer.wrap(primitive);
-        long firstLong = bb.getLong();
-        long secondLong = bb.getLong();
-        return new UUID(firstLong, secondLong);
-    }
-}
